connection tests passing  using picocontainer
