+++ src/cucumber/java/io.deepstream/RunCukesTest.java
@@ -1,0 +1,10 @@
+package io.deepstream;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions( features = "src/cucumber/resources")
+public class RunCukesTest {
+}
