+++ src/cucumber/resources/features.connecting/connecting.feature
@@ -5,9 +5,6 @@ Feature: Connecting a client
 
   Scenario: The test server is idle and awaits connections
     Given the test server is ready
-    Then the server has 0 active connections
-
-  #Scenario: The client is instantiated and creates a tcp connection
     And the client is initialised
     Then the server has 1 active connections
     And the clients connection state is "AWAITING_CONNECTION"
